ML_Server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
